+++ R/bmdplotwithgradient.R
@@ -9,7 +9,7 @@ bmdplotwithgradient <- function(extendedres, BMDtype = c("zSD", "xfold"),
                                 ncol4faceting, limits4colgradient,
                                 lowercol = "darkblue", uppercol = "darkred",
                                 add.label = FALSE, label.size = 2,
-                                BMD_log_transfo = FALSE)
+                                BMD_log_transfo = TRUE)
 {
   BMDtype <- match.arg(BMDtype, c("zSD", "xfold"))